feat: replace Agent[F] with typed Agent[F, In, Out]
